added albums and users routes
