+++ src/app.ts
@@ -1,9 +1,13 @@
 import * as express from 'express';
+import * as path from 'path';
+//import bodyParser from 'body-parser';
+var app=express();
 
-let app = express();
-app.get("/", (req, res)=>{
-  res.send("hello from NodeJS/Express");
-});
+//configure app
+//app.set('view engine','ejs');
+//app.set('views',path.join(__dirname,'views'));
 
-//start the  express application (web server) and keep listening for any reqest sent to port 1337
-app.listen(1337 , () => console.log("ready on port 1337"));
+app.use(require('./albums'));
+app.use(require('./users'));
+
+app.listen(1337, () => console.log('ready on port 1337') );
